+++ src/RayTree.Core/Tracking/ChangeTrackingBuilder.cs
@@ -2,7 +2,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using RayTree.Core.Distribution;
 using RayTree.Core.Handling;
-using RayTree.Core.Plugins;
+using RayTree.Core.Plugins.Compression;
 using RayTree.Core.Plugins.Deduplication;
 using RayTree.Core.Plugins.Outbox;
 using RayTree.Core.Plugins.Publisher;