+++ src/RayTree.Core/Tracking/EntityBuilder.cs
@@ -1,7 +1,7 @@
 using Microsoft.Extensions.Logging;
 using RayTree.Core.Distribution;
 using RayTree.Core.Handling;
-using RayTree.Core.Plugins;
+using RayTree.Core.Plugins.Compression;
 using RayTree.Core.Plugins.Consumer;
 using RayTree.Core.Plugins.Outbox;
 using RayTree.Core.Plugins.Publisher;