+++ src/RayTree.Plugins.Compressors.Brotli/BrotliCompressorPlugin.cs
@@ -1,5 +1,5 @@
 using System.IO.Compression;
-using RayTree.Core.Plugins;
+using RayTree.Core.Plugins.Compression;
 
 namespace RayTree.Plugins.Compressors.Brotli;