+++ src/RayTree.Plugins.Compressors.Gzip/GzipBuilderExtensions.cs
@@ -1,4 +1,4 @@
-using RayTree.Core.Plugins;
+using RayTree.Core.Plugins.Compression;
 using RayTree.Core.Tracking;
 
 namespace RayTree.Plugins.Compressors.Gzip;