+++ src/RayTree.Plugins.Compressors.Lz4/Lz4CompressorPlugin.cs
@@ -1,5 +1,5 @@
 using K4os.Compression.LZ4;
-using RayTree.Core.Plugins;
+using RayTree.Core.Plugins.Compression;
 
 namespace RayTree.Plugins.Compressors.Lz4;