+++ src/RayTree.Plugins.PostgreSQL/Outbox/Notification/NotificationBasedPublisher.cs
@@ -6,6 +6,7 @@ using Npgsql;
 using RayTree.Core.Distribution;
 using RayTree.Core.Models;
 using RayTree.Core.Plugins;
+using RayTree.Core.Plugins.Compression;
 using RayTree.Core.Plugins.Outbox;
 using RayTree.Core.Plugins.Publisher;
 using RayTree.Core.Plugins.Serialization;